+++ mathematica/unexpected-polynomial-identities-classical-interpolation.m
@@ -19,12 +19,19 @@ OddPowerForwardDecompositionMMinus1::usage="Validates the identity Odd power for
 OddPowerForwardDecompositionMMinus1Shifted::usage="Validates the identity Odd power forward decomposition m-1 shifted."
 OddPowerForwardDecompositionShifted::usage="Validates the identity Odd power forward decomposition shifted."
 OddPowerCentralDecomposition::usage="Validates the identity Odd power central decomposition."
-OddPowerBinomialForm::usage="Validates the identity Odd power binomial form shifted."
-OddPowerBinomialFormShifted::usage="Validates the identity Odd power binomial form."
 TableFormBackwardRecurrenceForT::usage="Prints the backward recurrence for Tm in the form of triangle."
 TableFormForwardRecurrenceForT::usage="Prints the forward recurrence for Tm in the form of triangle."
 TableFormCentralRecurrenceForT::usage="Prints the central recurrence for Tm in the form of triangle."
 TableFormBivariateSumT::usage="Prints Tm in the form of triangle."
+
+BinomialForm::usage="Validates the identity Binomial form."
+ShiftedBinomialForm::usage="Validates the identity Shifted binomial form."
+CenteredBinomialForm::usage="Validates the identity Centered binomial form."
+ShiftedCenteredBinomialForm::usage="Validates the identity Shifted centered binomial form."
+NegatedBinomialForm::usage="Validates the identity Negated binomial form."
+ShiftedNegatedBinomialForm::usage="Validates the identity Shifted negated binomial form."
+CenteredNegatedBinomialForm::usage="Validates the identity Centered negated binomial form."
+ShiftedCenteredNegatedBinomialForm::usage="Validates the identity Shifted centered negated binomial form."
 
 Begin["`Private`"]
 Unprotect[Power];
@@ -61,9 +68,14 @@ CentralRecurrenceForT[m_, n_, k_]:= Sum[(-1)^(t+1) * Binomial[m+1, t] * Bivariat
 TableFormCentralRecurrenceForT[m_, rows_]:= TableForm[Table[CentralRecurrenceForT[m, n, k], {n, -m/2, rows}, {k, 0, n+m/2}], TableAlignments -> Left];
 OddPowerCentralDecomposition[n_, m_]:= Sum[Sum[(-1)^(t+1) * Binomial[m+1, t] * BivariateSumT[m, n+(m/2)-t, k], {t, 1, m+1}], {k, 1, n+(m/2)}];
 
-OddPowerBinomialForm[m_, n_, a_] := Sum[A[m, r]* Sum[(k - a)^r * (n - a - k)^r, {k, a+1, n - a}], {r, 0, m}];
-OddPowerBinomialFormShifted[m_, n_, a_] := Sum[A[m, r]* Sum[(k - a)^r * (n - a - k)^r, {k, a, n - a - 1}], {r, 0, m}];
-TableFormBivariateSumT[m_, rows_]:=TableForm[Table[BivariateSumT[m, n, k], {n, 0, rows}, {k, 0, n}], TableAlignments -> Left];
+BinomialForm[m_, n_, a_] := Sum[A[m, r]* Sum[(k + a)^r * (n + a - k)^r, {k, -a+1, n + a}], {r, 0, m}];
+ShiftedBinomialForm[m_, n_, a_] := Sum[A[m, r]* Sum[(k + a)^r * (n + a - k)^r, {k, -a, n + a-1}], {r, 0, m}];
+CenteredBinomialForm[m_, n_, a_] := Sum[A[m, r]* Sum[(k + a/2)^r * (n + a/2 - k)^r, {k, -a/2 + 1, n + a/2}], {r, 0, m}];
+ShiftedCenteredBinomialForm[m_, n_, a_] := Sum[A[m, r]* Sum[(k + a/2)^r * (n + a/2 - k)^r, {k, -a/2, n + a/2 - 1}], {r, 0, m}];
+NegatedBinomialForm[m_, n_, a_] := Sum[A[m, r]* Sum[(k - a)^r * (n - a - k)^r, {k, a+1, n - a}], {r, 0, m}];
+ShiftedNegatedBinomialForm[m_, n_, a_] := Sum[A[m, r]* Sum[(k - a)^r * (n - a - k)^r, {k, a, n - a - 1}], {r, 0, m}];
+CenteredNegatedBinomialForm[m_, n_, a_] := Sum[A[m, r]* Sum[(k - a/2)^r * (n - a/2 - k)^r, {k, a/2, n - a/2 - 1}], {r, 0, m}];
+ShiftedCenteredNegatedBinomialForm[m_, n_, a_] := Sum[A[m, r]* Sum[(k - a/2)^r * (n - a/2 - k)^r, {k, a/2+1, n - a/2}], {r, 0, m}];
 
 End[ ]
 EndPackage[ ]